requirement.txt
python-dotenv
Pillow
customtkinter